requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
adafruit-blinka==5.9.1    # via adafruit-circuitpython-neopixel, adafruit-circuitpython-pypixelbuf
adafruit-circuitpython-neopixel==6.0.0  # via -r requirements.in
adafruit-circuitpython-pypixelbuf==2.2.0  # via adafruit-circuitpython-neopixel
adafruit-platformdetect==2.23.0  # via adafruit-blinka
adafruit-pureio==1.1.8    # via adafruit-blinka
pyftdi==0.52.0            # via adafruit-blinka
pyserial==3.5             # via pyftdi
pyusb==1.1.0              # via pyftdi
pyyaml==5.3.1             # via -r requirements.in
rpi-ws281x==4.2.5         # via -r requirements.in
